Fix many errors and bugs

+++ src/ThreeEditor/js/menubar/Menubar.File.js
@@ -1,14 +1,13 @@
-import { UIHorizontalRule, UIPanel, UIRow } from '../libs/ui.js';
+import { UIHorizontalRule, UIPanel } from '../libs/ui.js';
+import { createOption } from './Menubar.js';
 
 function MenubarFile(editor) {
-	const strings = editor.strings;
-
 	const container = new UIPanel();
 	container.setClass('menu');
 
 	const title = new UIPanel();
 	title.setClass('title');
-	title.setTextContent(strings.getKey('menubar/file'));
+	title.setTextContent('File');
 	container.add(title);
 
 	const options = new UIPanel();
@@ -17,22 +16,14 @@ function MenubarFile(editor) {
 
 	// New
 
-	let option = new UIRow();
-	option.setClass('option');
-	option.setTextContent(strings.getKey('menubar/file/new'));
-	option.onClick(() => {
-		if (window.confirm('Any unsaved data will be lost. Are you sure?')) {
-			editor.clear();
-		}
-	});
-	options.add(option);
-
-	//
-
-	options.add(new UIHorizontalRule());
+	options.add(
+		createOption('option', 'New', () => {
+			window.confirm('Any unsaved data will be lost. Are you sure?') && editor.clear();
+		}),
+		new UIHorizontalRule()
+	);
 
 	// Open Editor from file
-
 	const form = document.createElement('form');
 	form.style.display = 'none';
 	document.body.appendChild(form);
@@ -46,43 +37,14 @@ function MenubarFile(editor) {
 	});
 	form.appendChild(fileInput);
 
-	option = new UIRow();
-	option.setClass('option');
-	option.setTextContent('Open');
-	option.onClick(() => {
-		fileInput.click();
-	});
-	options.add(option);
-
-	//
-
-	options.add(new UIHorizontalRule());
+	options.add(
+		createOption('option', 'Open', () => {
+			fileInput.click();
+		}),
+		new UIHorizontalRule()
+	);
 
 	// Save Editor to file
-
-	option = new UIRow();
-	option.setClass('option');
-	option.setTextContent('Save');
-	option.onClick(() => {
-		editor.updateUserData();
-
-		let output = editor.toJSON();
-
-		try {
-			output = JSON.stringify(output, null, '\t');
-			output = output.replace(/[\n\t]+([\d.e\-[\]]+)/g, '$1');
-		} catch (e) {
-			output = JSON.stringify(output);
-		}
-
-		const fileName = window.prompt('Name of the file', 'editor');
-
-		if (fileName) saveString(output, `${fileName}.json`);
-	});
-	options.add(option);
-
-	//
-
 	const link = document.createElement('a');
 	function save(blob, filename) {
 		if (link.href) {
@@ -93,10 +55,28 @@ function MenubarFile(editor) {
 		link.download = filename || 'data.json';
 		link.dispatchEvent(new MouseEvent('click'));
 	}
-
 	function saveString(text, filename) {
 		save(new Blob([text], { type: 'text/plain' }), filename);
 	}
+
+	options.add(
+		createOption('option', 'Save', () => {
+			editor.updateUserData();
+
+			let output = editor.toJSON();
+
+			try {
+				output = JSON.stringify(output, null, '\t');
+				output = output.replace(/[\n\t]+([\d.e\-[\]]+)/g, '$1');
+			} catch (e) {
+				output = JSON.stringify(output);
+			}
+
+			const fileName = window.prompt('Name of the file', 'editor');
+
+			if (fileName) saveString(output, `${fileName}.json`);
+		})
+	);
 
 	return container;
 }
